refactor : ResponseEntity 응답객체 MessageVO변경 및 id 자료형 Long으로 변경

+++ src/main/java/com/example/bsm/mapper/PersonalRecordMapper.java
@@ -15,11 +15,11 @@ public interface PersonalRecordMapper {
 
     int updatePersonalRecord(PersonalRecordVO personalRecordVO);
 
-    int deletePersonalRecord(int id);
+    int deletePersonalRecord(long id);
 
     // 기록 전체 조회
     List<PersonalRecordVO> getAllRecord(Pagination pagination);
 
     // 사용자 기록 전체 조회
-    List<PersonalRecordVO> getPersonalRecord(int memberId);
+    List<PersonalRecordVO> getPersonalRecord(long memberId);
 }
